alden-be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-genai==0.1.0
opencv-python==4.8.1.78
pyaudio==0.2.11
pillow==10.1.0
mss==9.0.1
python-dotenv==1.0.0 